+++ src/Solvers/L2/Solver.h
@@ -13,41 +13,41 @@ namespace L2Solver {
         private:
             digraph<int> clone_tree;
             std::unordered_map<int, int> vertex_map;
-            std::vector<std::vector<double>> frequency_matrix;
-            std::vector<std::vector<double>> weight_matrix;
+            std::vector<std::vector<float>> frequency_matrix;
+            std::vector<std::vector<float>> weight_matrix;
             int root;
         
             std::unordered_map<int, std::vector<PiecewiseQuadraticF>> fs;
             std::unordered_map<int, std::vector<PiecewiseQuadraticF>> gs;
-            std::vector<std::vector<double>> alphas;
+            std::vector<std::vector<float>> alphas;
 
             /* Backtracks to compute the dual 
              * variables for sample j */
-            void backtrack(double alpha_0, int j);
+            void backtrack(float alpha_0, int j);
         public:
-            double objective;
-            std::vector<std::vector<double>> frequencies;
+            float objective;
+            std::vector<std::vector<float>> frequencies;
 
             Solver(
                 digraph<int> clone_tree, 
                 std::unordered_map<int, int> vertex_map, 
-                std::vector<std::vector<double>> frequency_matrix, 
-                std::vector<std::vector<double>> weight_matrix,
+                std::vector<std::vector<float>> frequency_matrix, 
+                std::vector<std::vector<float>> weight_matrix,
                 int root
             ) : clone_tree(clone_tree), vertex_map(vertex_map), root(root) {
                 this->frequency_matrix.reserve(frequency_matrix.size());
                 for (size_t i = 0; i < frequency_matrix.size(); i++) {
-                    this->frequency_matrix.push_back(std::vector<double>(frequency_matrix[i].begin(), frequency_matrix[i].end()));
+                    this->frequency_matrix.push_back(std::vector<float>(frequency_matrix[i].begin(), frequency_matrix[i].end()));
                 }
 
                 for (size_t i = 0; i < frequency_matrix.size(); i++) {
-                    this->alphas.push_back(std::vector<double>(frequency_matrix[i].size()));
-                    this->frequencies.push_back(std::vector<double>(frequency_matrix[i].size()));
+                    this->alphas.push_back(std::vector<float>(frequency_matrix[i].size()));
+                    this->frequencies.push_back(std::vector<float>(frequency_matrix[i].size()));
                 }
 
                 this->weight_matrix.reserve(weight_matrix.size());
                 for (size_t i = 0; i < weight_matrix.size(); i++) {
-                    this->weight_matrix.push_back(std::vector<double>(weight_matrix[i].begin(), weight_matrix[i].end()));
+                    this->weight_matrix.push_back(std::vector<float>(weight_matrix[i].begin(), weight_matrix[i].end()));
                 }
             }
